requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
redis==3.5.3
pysocks==1.7.1
ipaddress==1.0.23
pyyaml==6.0
concurrent-futures==3.0.5 